+++ src/Test/intlinear.jl
@@ -54,6 +54,9 @@ function int1test(model::MOI.ModelLike, config::TestConfig)
     @test MOI.get(model, MOI.ObjectiveSense()) == MOI.MaxSense
 
     if config.solve
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
         MOI.optimize!(model)
 
         @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -123,6 +126,9 @@ function int2test(model::MOI.ModelLike, config::TestConfig)
         @test MOI.get(model, MOI.ObjectiveSense()) == MOI.MaxSense
 
         if config.solve
+            @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+            @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
             MOI.optimize!(model)
 
             @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -209,6 +215,9 @@ function int2test(model::MOI.ModelLike, config::TestConfig)
         @test MOI.get(model, MOI.ObjectiveSense()) == MOI.MaxSense
 
         if config.solve
+            @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+            @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
             MOI.optimize!(model)
 
             @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -279,6 +288,9 @@ function int3test(model::MOI.ModelLike, config::TestConfig)
     MOI.set(model, MOI.ObjectiveSense(), MOI.MaxSense)
 
     if config.solve
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
         MOI.optimize!(model)
 
         @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success
@@ -336,6 +348,9 @@ function knapsacktest(model::MOI.ModelLike, config::TestConfig)
     end
 
     if config.solve
+        @test MOI.get(model, MOI.TerminationStatus()) == MOI.OptimizeNotCalled
+        @test MOI.get(model, MOI.PrimalStatus()) == MOI.NoSolution
+
         MOI.optimize!(model)
 
         @test MOI.get(model, MOI.TerminationStatus()) == MOI.Success